+++ custom_components/soundtouchplus/config_flow.py
@@ -19,13 +19,15 @@ from __future__ import annotations
 
 import logging
 
-from bosesoundtouchapi import SoundTouchDevice
+from bosesoundtouchapi import SoundTouchDevice, SoundTouchClient
+from bosesoundtouchapi.models import SourceList, SourceItem
 from requests import RequestException
 from typing import Any
 import voluptuous as vol
 
 from homeassistant import config_entries, core, exceptions
 from homeassistant.components import zeroconf
+from homeassistant.config_entries import ConfigEntry, ConfigFlow, OptionsFlow
 from homeassistant.const import (
     CONF_HOST,
     CONF_NAME,
@@ -40,6 +42,7 @@ from .const import (
     DOMAIN,
     CONF_DEVICE_NAME,
     CONF_DEVICE_ID,
+    CONF_OPTION_SOURCE_LIST,
     CONF_PING_WEBSOCKET_INTERVAL,
     CONF_PORT_WEBSOCKET,
     DEFAULT_PING_WEBSOCKET_INTERVAL,
@@ -80,9 +83,7 @@ async def validate_device_connection_http(hass:HomeAssistant, data:dict) -> dict
         # create new instance of SoundTouchDevice to ensure we can connect via HTTP.
         # we have to use the executor since the bosesoundtouchapi does not support async.
         # SoundTouchDevice init arguments: host, connectTimeout, proxyManager, port
-        device:SoundTouchDevice = await hass.async_add_executor_job(
-             SoundTouchDevice, host, timeout, None, port
-        )
+        device:SoundTouchDevice = await hass.async_add_executor_job(SoundTouchDevice, host, timeout, None, port)
 
         _logsi.LogVerbose("ConfigFlow HTTP connection is valid for device: IP=%s, port=%s, name=%s, id=%s" % (host, port, device.DeviceName, device.DeviceId))
 
@@ -121,6 +122,18 @@ class SoundTouchPlusConfigFlow(config_entries.ConfigFlow, domain=DOMAIN):
         self._ping_websocket_interval: int | None = DEFAULT_PING_WEBSOCKET_INTERVAL
         self._port: int | None = DEFAULT_PORT
         self._port_websocket: int | None = DEFAULT_PORT_WEBSOCKET
+
+
+    @staticmethod
+    @callback
+    def async_get_options_flow(config_entry: ConfigEntry) -> OptionsFlow:
+        """
+        Get the options flow for this handler, which enables options support.
+
+        This method is invoked when a user clicks the "Configure" button from the integration
+        details page of the UI.
+        """
+        return SoundTouchPlusOptionsFlow(config_entry)
 
 
     async def async_step_user(self, user_input=None):
@@ -301,7 +314,7 @@ class SoundTouchPlusConfigFlow(config_entries.ConfigFlow, domain=DOMAIN):
         configData:dict = self._get_data()
 
         # create the configuration entry.
-        _logsi.LogDictionary(SILevel.Verbose, "ConfigFlow is creating a configuration entry for a deviceIP=%s, port=%s, name=%s, id=%s" % (self._host, self._port, self._name, self._device_id), configData)
+        _logsi.LogDictionary(SILevel.Verbose, "ConfigFlow is creating a configuration entry for deviceIP=%s, port=%s, name=%s, id=%s" % (self._host, self._port, self._name, self._device_id), configData)
         return self.async_create_entry(
             title=self._name or self._host,
             data=configData,
@@ -337,3 +350,141 @@ class InvalidAuth(exceptions.HomeAssistantError):
 
 class WSCannotConnect(exceptions.HomeAssistantError):
     """Error to indicate we cannot connect to websocket."""
+
+
+class SoundTouchPlusOptionsFlow(OptionsFlow):
+    """
+    Handles options flow for the component.
+    
+    The options flow allows a user to configure additional options for the component at any time by 
+    navigating to the integrations page and clicking the Options button on the card for your component. 
+    Generally speaking these configuration values are optional, whereas values in the config flow are 
+    required to make the component function.
+    """
+
+    def __init__(self, configEntry:ConfigEntry) -> None:
+        """
+        Initialize options flow.
+        """
+        if _logsi.IsOn(SILevel.Verbose):
+            _logsi.LogVerbose("OptionsFlow is initializing")
+            _logsi.LogObject(SILevel.Verbose, "OptionsFlow config_entry object", configEntry)
+            _logsi.LogDictionary(SILevel.Verbose, "OptionsFlow config_entry.data dictionary", configEntry.data)
+            _logsi.LogDictionary(SILevel.Verbose, "OptionsFlow config_entry.options dictionary", configEntry.options)
+
+        # initialize storage.
+        self._name:str = None
+        self._host:str = None
+        self._port:int = DEFAULT_PORT
+
+        # always check for keys, in case of an upgrade that contains a new key
+        # that is not present in a previous version.           
+
+        # load config entry base values.
+        self._ConfigEntry = configEntry
+        if CONF_NAME in configEntry.data.keys():
+            self._name = configEntry.data[CONF_NAME]
+        if CONF_HOST in configEntry.data.keys():
+            self._host = configEntry.data[CONF_HOST]
+        if CONF_PORT in configEntry.data.keys():
+            self._port = configEntry.data[CONF_PORT]
+
+        # load config entry options values.
+        self._Options = dict(configEntry.options)
+
+
+    async def async_step_init(self, user_input:dict[str,Any]=None) -> FlowResult:
+        """
+        Manage the options for the custom component.
+        
+        For a good example, look at HA demo source code:
+            \home-assistant-core\homeassistant\components\demo\config_flow.py
+        """
+        errors: dict[str, str] = {}
+        
+        # the user_input variable defaults to None when this step is first called. 
+        # when the user clicks the submit button on the form, the user_input variable will be 
+        # a dictionary containing the data that was entered.  Home Assistant will do some basic 
+        # validation on your behalf based on the data schema that you defined (e.g. required field,
+        # port number is within a numeric range, etc). 
+        if user_input is not None:
+            
+            # sort updated config entry options.
+            if CONF_OPTION_SOURCE_LIST in user_input:
+                user_input[CONF_OPTION_SOURCE_LIST].sort()
+
+            # store the updated config entry options.
+            self._Options.update(user_input)
+            return await self._update_options()
+
+        # load available sources from the device.
+        source_list_all:dict = await self.hass.async_add_executor_job(self._GetSourceTitleList, self._host, self._port, self._name)
+        if source_list_all is None:
+            errors["base"] = "getsourcelist_empty"
+            return
+
+        # log sources that are currently selected.
+        source_list_selected:list[str] = self._Options.get(CONF_OPTION_SOURCE_LIST, [])
+        _logsi.LogArray(SILevel.Verbose, "(%s): %s - OptionsFlow currently SELECTED '%s' items for device" % (self._host, self._name, CONF_OPTION_SOURCE_LIST), source_list_selected)
+                   
+        # create validation schema; default source list to empty array, which will cause the
+        schema = vol.Schema(
+            {
+                vol.Optional(CONF_OPTION_SOURCE_LIST, default=self._Options.get(CONF_OPTION_SOURCE_LIST, [])): cv.multi_select(source_list_all),
+            }
+        )
+
+        _logsi.LogVerbose("(%s): %s - OptionsFlow is showing the init configuration options form" % (self._host, self._name))
+        return self.async_show_form(
+            step_id="init", 
+            data_schema=schema, 
+            description_placeholders={CONF_NAME: self._name},
+            errors=errors or {}
+        )
+    
+    
+    async def _update_options(self) -> FlowResult:
+        """
+        Update config entry options.
+        """
+        _logsi.LogDictionary(SILevel.Verbose, "(%s): %s - OptionsFlow UPDATED options dictionary for device" % (self._host, self._name), self._Options)
+
+        return self.async_create_entry(
+            title="", 
+            data=self._Options
+        )
+    
+
+    def _GetSourceTitleList(self, host:str, port:int, name:str) -> dict:
+        """
+        Retrieves SourceList object from the SoundTouch device.
+        """
+        try:
+
+            _logsi.LogVerbose("(%s): %s - OptionsFlow is creating SoundTouchDevice instance (port=%s)" % (host, name, port))
+            device:SoundTouchDevice = SoundTouchDevice(host, 30, None, port)
+
+            _logsi.LogVerbose("(%s): %s - OptionsFlow is creating SoundTouchClient instance" % (host, name))
+            client:SoundTouchClient = SoundTouchClient(device)
+
+            _logsi.LogVerbose("(%s): %s - OptionsFlow is retrieving SourceList configuration for device" % (host, name))
+            sourceList:SourceList = client.GetSourceList()
+            
+            # sort the results (in place) by SourceTitle, ascending order.
+            sourceList.SourceItems.sort(key=lambda x: (x.SourceTitle or "").lower(), reverse=False)
+           
+            # build string array of all sources for the device.
+            result:dict = {}
+            item:SourceItem
+            for item in sourceList:
+                if item.SourceTitle not in result:
+                    result[item.SourceTitle] = item.SourceTitle
+
+            _logsi.LogDictionary(SILevel.Verbose, "(%s): %s - OptionsFlow ALL available '%s' items for device" % (host, name, CONF_OPTION_SOURCE_LIST), result)
+            
+            return result
+            
+        except Exception as ex:
+            
+            _logsi.LogError("(%s): %s - OptionsFlow could not retrieve source list for device: %s" % (host, name, str(ex)))
+            return None